#2 Pull request - Update de auditoria y cosas de adoptante y usuario

+++ src/main/java/co/edu/unbosque/veterinaria/repository/AplicacionVacunaRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AplicacionVacunaRepository extends CrudRepository<AplicacionVacuna, Long> {
-}
+public interface AplicacionVacunaRepository extends CrudRepository<AplicacionVacuna, Integer> { }
