+++ src/main/java/co/edu/unbosque/veterinaria/repository/VeterinarioRepository.java
@@ -5,5 +5,4 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VeterinarioRepository extends CrudRepository<Veterinario, Long> {
-}
+public interface VeterinarioRepository extends CrudRepository<Veterinario, Integer> { }
