+++ src/main/java/co/edu/unbosque/veterinaria/service/api/AdopcionServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Adopcion;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface AdopcionServiceAPI extends GenericServiceAPI<Adopcion, Long> { }
+public interface AdopcionServiceAPI extends GenericServiceAPI<Adopcion, Integer> { }
