+++ src/main/java/co/edu/unbosque/veterinaria/service/api/AplicacionVacunaServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.AplicacionVacuna;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface AplicacionVacunaServiceAPI extends GenericServiceAPI<AplicacionVacuna, Long> { }
+public interface AplicacionVacunaServiceAPI extends GenericServiceAPI<AplicacionVacuna, Integer> { }
