+++ src/main/java/co/edu/unbosque/veterinaria/service/api/AuditoriaServiceAPI.java
@@ -3,4 +3,5 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Auditoria;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface AuditoriaServiceAPI extends GenericServiceAPI<Auditoria, Long> { }
+public interface AuditoriaServiceAPI extends GenericServiceAPI<Auditoria, Integer> {
+}
