+++ src/main/java/co/edu/unbosque/veterinaria/service/api/CuidadorServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Cuidador;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface CuidadorServiceAPI extends GenericServiceAPI<Cuidador, Long> { }
+public interface CuidadorServiceAPI extends GenericServiceAPI<Cuidador, Integer> { }
