+++ src/main/java/co/edu/unbosque/veterinaria/service/api/DiagnosticoServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Diagnostico;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface DiagnosticoServiceAPI extends GenericServiceAPI<Diagnostico, Long> { }
+public interface DiagnosticoServiceAPI extends GenericServiceAPI<Diagnostico, Integer> { }
