+++ src/main/java/co/edu/unbosque/veterinaria/service/api/EmpleadoServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Empleado;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface EmpleadoServiceAPI extends GenericServiceAPI<Empleado, Long> { }
+public interface EmpleadoServiceAPI extends GenericServiceAPI<Empleado, Integer> { }
