+++ src/main/java/co/edu/unbosque/veterinaria/service/api/RazaServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Raza;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface RazaServiceAPI extends GenericServiceAPI<Raza, Long> { }
+public interface RazaServiceAPI extends GenericServiceAPI<Raza, Integer> { }
