+++ src/main/java/co/edu/unbosque/veterinaria/service/api/VeterinarioServiceAPI.java
@@ -3,4 +3,4 @@ package co.edu.unbosque.veterinaria.service.api;
 import co.edu.unbosque.veterinaria.entity.Veterinario;
 import co.edu.unbosque.veterinaria.utils.GenericServiceAPI;
 
-public interface VeterinarioServiceAPI extends GenericServiceAPI<Veterinario, Long> { }
+public interface VeterinarioServiceAPI extends GenericServiceAPI<Veterinario, Integer> { }
